tidy up auth backends
