+++ backend/api/auth_backends/__init__.py
@@ -1,5 +1,5 @@
-from .dependent_student import (
-    DependentStudentBackend,
-    DependentStudentTokenBackend,
+from .email_and_password import EmailAndPasswordBackend
+from .user_id_and_token import UserIdAndTokenBackend
+from .username_and_password_and_class_id import (
+    UsernameAndPasswordAndClassIdBackend,
 )
-from .teacher_or_independent_student import TeacherOrIndependentStudentBackend
