+++ backend/api/auth_backends/email_and_password.py
@@ -8,15 +8,25 @@ from django.http.request import HttpRequest
 User = get_user_model()
 
 
-class TeacherOrIndependentStudentBackend(BaseBackend):
+class EmailAndPasswordBackend(BaseBackend):
     def authenticate(
         self,
         request: HttpRequest,
-        username: t.Optional[str] = None,
+        email: t.Optional[str] = None,
         password: t.Optional[str] = None,
         **kwargs
     ) -> t.Optional[AbstractBaseUser]:
-        return super().authenticate(request, username, password, **kwargs)
+        if email is None or password is None:
+            return
+
+        try:
+            user = User.objects.get(email=email)
+            if getattr(user, "is_active", True) and user.check_password(
+                password
+            ):
+                return user
+        except User.DoesNotExist:
+            return
 
     def get_user(self, user_id: int) -> t.Optional[AbstractBaseUser]:
         try:
